data-replay/requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
requests
